pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "mnemosyne-orchestration"
version = "1.0.0"
description = "Multi-agent orchestration system for Mnemosyne with PyO3 bindings"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.72.0",
    "claude-agent-sdk>=0.1.0",
    "dspy-ai>=3.0.3",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "maturin>=1.0",
]

[tool.maturin]
module-name = "mnemosyne_core"
bindings = "pyo3"
features = ["python"]

[tool.pytest.ini_options]
markers = [
    "integration: Integration tests that require ANTHROPIC_API_KEY (deselect with '-m \"not integration\"')",
    "slow: Slow-running tests that can be skipped for quick validation",
]
asyncio_mode = "auto"
